+++ dash/_pages.py
@@ -9,13 +9,11 @@ from os.path import isfile, join
 from pathlib import Path
 from urllib.parse import parse_qs
 
-import flask
-
 from . import _validate
 from ._callback_context import context_value
 from ._get_app import get_app
 from ._get_paths import get_relative_path
-from ._utils import AttributeDict
+from ._utils import AttributeDict, get_root_path
 
 CONFIG = AttributeDict()
 PAGE_REGISTRY = collections.OrderedDict()
@@ -98,7 +96,7 @@ def _path_to_module_name(path):
 def _infer_module_name(page_path):
     relative_path = page_path.split(CONFIG.pages_folder)[-1]
     module = _path_to_module_name(relative_path)
-    proj_root = flask.helpers.get_root_path(CONFIG.name)
+    proj_root = get_root_path(CONFIG.name)
     if CONFIG.pages_folder.startswith(proj_root):
         parent_path = CONFIG.pages_folder[len(proj_root) :]
     else:
@@ -150,23 +148,12 @@ def _parse_path_variables(pathname, path_template):
     return dict(zip(var_names, variables))
 
 
-def _create_redirect_function(redirect_to):
-    def redirect():
-        return flask.redirect(redirect_to, code=301)
-
-    return redirect
-
-
 def _set_redirect(redirect_from, path):
     app = get_app()
     if redirect_from and len(redirect_from):
         for redirect in redirect_from:
             fullname = app.get_relative_path(redirect)
-            app.server.add_url_rule(
-                fullname,
-                fullname,
-                _create_redirect_function(app.get_relative_path(path)),
-            )
+            app.backend.add_redirect_rule(app, fullname, app.get_relative_path(path))
 
 
 def register_page(
@@ -318,18 +305,22 @@ def register_page(
     )
     page.update(
         supplied_title=title,
-        title=title
-        if title is not None
-        else CONFIG.title
-        if CONFIG.title != "Dash"
-        else page["name"],
+        title=(
+            title
+            if title is not None
+            else CONFIG.title
+            if CONFIG.title != "Dash"
+            else page["name"]
+        ),
     )
     page.update(
-        description=description
-        if description
-        else CONFIG.description
-        if CONFIG.description
-        else "",
+        description=(
+            description
+            if description
+            else CONFIG.description
+            if CONFIG.description
+            else ""
+        ),
         order=order,
         supplied_order=order,
         supplied_layout=layout,
@@ -389,16 +380,14 @@ def _path_to_page(path_id):
     return {}, None
 
 
-def _page_meta_tags(app):
-    start_page, path_variables = _path_to_page(flask.request.path.strip("/"))
+def _page_meta_tags(app, request):
+    request_path = request.path
+    start_page, path_variables = _path_to_page(request_path.strip("/"))
 
-    # use the supplied image_url or create url based on image in the assets folder
     image = start_page.get("image", "")
     if image:
         image = app.get_asset_url(image)
-    assets_image_url = (
-        "".join([flask.request.url_root, image.lstrip("/")]) if image else None
-    )
+    assets_image_url = "".join([request.root, image.lstrip("/")]) if image else None
     supplied_image_url = start_page.get("image_url")
     image_url = supplied_image_url if supplied_image_url else assets_image_url
 
@@ -413,7 +402,7 @@ def _page_meta_tags(app):
     return [
         {"name": "description", "content": description},
         {"property": "twitter:card", "content": "summary_large_image"},
-        {"property": "twitter:url", "content": flask.request.url},
+        {"property": "twitter:url", "content": request.url},
         {"property": "twitter:title", "content": title},
         {"property": "twitter:description", "content": description},
         {"property": "twitter:image", "content": image_url or ""},